fix: rounding errors for tax-included items
This fix adds a new method for adjusting the calculated inclusive tax for
items if needed so that inclusive taxes + net amount = inclusive item rate.

For example: if an inclusive rate is 35 with an included VAT of 15%:
* Calculated net rate is 35 / 1.15 = 30.4347 ~= 30.43
* VAT on calculated net rate is 30.43 * .15 = 4.5645 ~= 4.56
* The sum is 30.43 + 4.56 = 34.99

This causes problems with VAT authorities that require taxes to add up (at the
item and/or invoice level). The new method detects that the sum of inclusive
taxes should be (amount - net_amount), i.e. (35 - 30.43 = 4.57) in the above
case, and adjusts the tax accordingly by 0.01

The new method is controlled by a new check in `Accounts Settings` that
is turned on by default.

For UAE/SA, the `update_itemised_tax_data` hook was also updated to make
use of the adjusted tax values.